Add DepositStorage API's to StorageSystem
