+++ storage/build.gradle
@@ -22,6 +22,7 @@ dependencies {
   testImplementation testFixtures(project(':ethereum:core'))
   testImplementation testFixtures(project(':ethereum:datastructures'))
   testImplementation testFixtures(project(':util'))
+  testImplementation testFixtures(project(':pow'))
   testImplementation testFixtures(project(':storage'))
 
   testFixturesApi project(':util')
@@ -32,6 +33,7 @@ dependencies {
   testFixturesImplementation testFixtures(project(':bls'))
   testFixturesImplementation testFixtures(project(':data:metrics'))
   testFixturesImplementation testFixtures(project(':ethereum:core'))
+  testFixturesImplementation testFixtures(project(':pow'))
   testFixturesImplementation project(':protoarray')
   testFixturesImplementation testFixtures(project(':util'))
   testFixturesImplementation 'com.google.guava:guava'
