+++ storage/src/testFixtures/java/tech/pegasys/teku/storage/storageSystem/FileBackedStorageSystem.java
@@ -17,6 +17,7 @@ import com.google.common.eventbus.EventBus;
 import java.nio.file.Path;
 import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import tech.pegasys.teku.metrics.StubMetricsSystem;
+import tech.pegasys.teku.pow.api.TrackingEth1EventsChannel;
 import tech.pegasys.teku.storage.api.FinalizedCheckpointChannel;
 import tech.pegasys.teku.storage.api.StubFinalizedCheckpointChannel;
 import tech.pegasys.teku.storage.api.TrackingReorgEventChannel;
@@ -25,6 +26,7 @@ import tech.pegasys.teku.storage.client.RecentChainData;
 import tech.pegasys.teku.storage.client.StorageBackedRecentChainData;
 import tech.pegasys.teku.storage.server.ChainStorage;
 import tech.pegasys.teku.storage.server.Database;
+import tech.pegasys.teku.storage.server.DepositStorage;
 import tech.pegasys.teku.storage.server.rocksdb.RocksDbConfiguration;
 import tech.pegasys.teku.storage.server.rocksdb.RocksDbDatabase;
 import tech.pegasys.teku.util.config.StateStorageMode;
@@ -34,6 +36,7 @@ public class FileBackedStorageSystem extends AbstractStorageSystem implements St
 
   private final EventBus eventBus;
   private final TrackingReorgEventChannel reorgEventChannel;
+  private final TrackingEth1EventsChannel eth1EventsChannel = new TrackingEth1EventsChannel();
 
   private final CombinedChainDataClient combinedChainDataClient;
   private final Database database;
@@ -98,6 +101,11 @@ public class FileBackedStorageSystem extends AbstractStorageSystem implements St
   }
 
   @Override
+  public DepositStorage createDepositStorage(final boolean eth1DepositsFromStorageEnabled) {
+    return DepositStorage.create(eth1EventsChannel, database, eth1DepositsFromStorageEnabled);
+  }
+
+  @Override
   public Database getDatabase() {
     return database;
   }
@@ -130,5 +138,10 @@ public class FileBackedStorageSystem extends AbstractStorageSystem implements St
   @Override
   public TrackingReorgEventChannel reorgEventChannel() {
     return reorgEventChannel;
+  }
+
+  @Override
+  public TrackingEth1EventsChannel eth1EventsChannel() {
+    return eth1EventsChannel;
   }
 }
